Essence rework, Numbers, PlayerName etc

+++ Source/SkyCraft/Damage.h
@@ -56,7 +56,7 @@ public:
 	UPROPERTY(BlueprintReadWrite, EditAnywhere) bool bIsPercentage = false;
 	UPROPERTY(BlueprintReadWrite, EditAnywhere) bool bIsLinearDamage = false;
 	UPROPERTY(BlueprintReadWrite, EditAnywhere, meta=(EditCondition="bIsLinearDamage", EditConditionHides)) int32 MaxLinearDamage = 0;
-	UPROPERTY(BlueprintReadWrite, EditAnywhere, meta=(EditCondition="bIsLinearDamage", EditConditionHides)) int32 MaxLinearHitMass = 0;
+	UPROPERTY(BlueprintReadWrite, EditAnywhere, meta=(EditCondition="bIsLinearDamage", EditConditionHides)) float MaxLinearHitMass = 0;
 	UPROPERTY(BlueprintReadWrite, EditAnywhere) float HitMass = 0;
 	UPROPERTY(BlueprintReadWrite, EditAnywhere) EDamageGlobalType DamageGlobalType = EDamageGlobalType::Pure;
 	UPROPERTY(BlueprintReadWrite, EditAnywhere) bool bShowDamageNumbers = true;
